Move Result and Error to Domain namspace
