+++ Catan/Catan.Core/Abstractions/IQuery.cs
@@ -1,7 +1,7 @@
-﻿using Catan.Core.Models;
+﻿using Catan.Domain;
 using MediatR;
 
-namespace Catan.Core;
+namespace Catan.Core.Abstractions;
 
 public interface IQuery : IRequest<Result>;
 public interface IQuery<TResponse> : IRequest<Result<TResponse>>;
