+++ Catan/Catan.Core/GameActions/BuildCity/BuildCityCommandHandler.cs
@@ -1,6 +1,6 @@
-﻿using Catan.Core.Models;
+﻿using Catan.Core.Abstractions;
 using Catan.Core.Services;
-using Catan.Domain.Enums;
+using Catan.Domain;
 
 namespace Catan.Core.GameActions.BuildCity;
