+++ Catan/Catan.Core/GameActions/BuildSettlement/BuildSettlementCommandHandler.cs
@@ -1,6 +1,7 @@
-﻿using Catan.Core.Models;
+﻿using Catan.Domain;
 using Catan.Core.Services;
 using Catan.Domain.Enums;
+using Catan.Core.Abstractions;
 
 namespace Catan.Core.GameActions.BuildSettlement;
