+++ Catan/Catan.Core/GameActions/BuyDevelopmentCard/BuyDevelopmentCardCommand.cs
@@ -1,3 +1,5 @@
-﻿namespace Catan.Core.GameActions.BuyDevelopmentCard;
+﻿using Catan.Core.Abstractions;
+
+namespace Catan.Core.GameActions.BuyDevelopmentCard;
 
 public sealed record BuyDevelopmentCardCommand(string GameId) : ICommand;
