+++ Catan/Catan.Core/GameActions/CreateGame/CreateGameCommand.cs
@@ -1,4 +1,6 @@
-﻿namespace Catan.Core.GameActions.CreateGame;
+﻿using Catan.Core.Abstractions;
+
+namespace Catan.Core.GameActions.CreateGame;
 
 public sealed record CreateGameCommand(
     int PlayerCount,
