+++ Catan/Catan.Core/GameActions/CreateGame/CreateGameCommandHandler.cs
@@ -1,4 +1,4 @@
-﻿using Catan.Core.Models;
+﻿using Catan.Core.Abstractions;
 using Catan.Core.Services;
 using Catan.Domain;
