+++ Catan/Catan.Core/GameActions/EmbargoPlayer/EmbargoPlayerCommand.cs
@@ -1,4 +1,6 @@
-﻿namespace Catan.Core.GameActions.EmbargoPlayer;
+﻿using Catan.Core.Abstractions;
+
+namespace Catan.Core.GameActions.EmbargoPlayer;
 
 public sealed record EmbargoPlayerCommand(
     string GameId,
