+++ Catan/Catan.Core/GameActions/EndTurn/EndTurnCommand.cs
@@ -1,3 +1,5 @@
-﻿namespace Catan.Core.GameActions.EndTurn;
+﻿using Catan.Core.Abstractions;
+
+namespace Catan.Core.GameActions.EndTurn;
 
 public sealed record EndTurnCommand(string GameId) : ICommand;
