+++ Catan/Catan.Core/GameActions/GetAvailableCityLocations/GetAvailableCityLocationsQueryHandler.cs
@@ -1,4 +1,5 @@
-﻿using Catan.Core.Models;
+﻿using Catan.Core.Abstractions;
+using Catan.Core.Models;
 using Catan.Core.Services;
 using Catan.Domain;
 using Catan.Domain.Enums;
