+++ Catan/Catan.Core/GameActions/GetGame/GetGameQueryHandler.cs
@@ -1,5 +1,7 @@
-﻿using Catan.Core.Models;
+﻿using Catan.Core.Abstractions;
+using Catan.Core.Models;
 using Catan.Core.Services;
+using Catan.Domain;
 using Catan.Domain.Enums;
 
 namespace Catan.Core.GameActions.GetGame;
