+++ Catan/Catan.Core/GameActions/MakeTradeOffer/MakeTradeOfferCommand.cs
@@ -1,4 +1,5 @@
-﻿using Catan.Domain.Enums;
+﻿using Catan.Core.Abstractions;
+using Catan.Domain.Enums;
 
 namespace Catan.Core.GameActions.MakeTradeOffer;
