+++ Catan/Catan.Core/GameActions/PlayMonopolyCard/PlayMonopolyCardCommandHandler.cs
@@ -1,5 +1,6 @@
-﻿using Catan.Core.Models;
+﻿using Catan.Core.Abstractions;
 using Catan.Core.Services;
+using Catan.Domain;
 using Catan.Domain.Enums;
 
 namespace Catan.Core.GameActions.PlayMonopolyCard;
