+++ Catan/Catan.Core/GameActions/RollDice/RollDiceCommand.cs
@@ -1,3 +1,5 @@
-﻿namespace Catan.Core.GameActions.RollDice;
+﻿using Catan.Core.Abstractions;
+
+namespace Catan.Core.GameActions.RollDice;
 
 public sealed record RollDiceCommand(string GameId) : ICommand<RollDiceResponse>;
