+++ Catan/Catan.Core/GameActions/TradeWithBank/TradeWithBankCommand.cs
@@ -1,4 +1,6 @@
-﻿namespace Catan.Core.GameActions.TradeWithBank;
+﻿using Catan.Core.Abstractions;
+
+namespace Catan.Core.GameActions.TradeWithBank;
 
 public sealed record TradeWithBankCommand(
     string GameId,
