+++ Catan/Catan.Domain/Errors.cs
@@ -1,6 +1,6 @@
 ﻿using System.Net;
 
-namespace Catan.Core.Models;
+namespace Catan.Domain;
 
 public static class Errors
 {
